float

+++ 01 c_structure.c
@@ -1,11 +1,11 @@
-// Write a program how to printout two user input
+// Write a program how to print out two user input
 int main()
 {
 
-    int num1,num2;
+    float num1,num2;
     printf("Enter two Integers: ");
-    scanf("%d %d",&num1,&num2);
-    printf("Number's are : %d ,%d",num1,num2);
+    scanf("%f %f",&num1,&num2);
+    printf("Number's are : %f ,%f",num1,num2);
 
     return 0;
